Reorganized some types.
