+++ packages/types/components/index.ts
@@ -5,7 +5,6 @@ export * from './Buttonifier'
 export * from './ButtonPopup'
 export * from './Collapsible'
 export * from './CommandModal'
-export * from './common'
 export * from './CopyToClipboard'
 export * from './DaoCard'
 export * from './DaoDappTabbedHome'
