+++ packages/types/gov.ts
@@ -17,9 +17,9 @@ import {
 } from '@dao-dao/utils/protobuf/codegen/cosmos/upgrade/v1beta1/upgrade'
 
 import { NestedActionsEditorFormData } from './actions'
-import { LoadingData } from './components'
 import { Coin, CosmosMsgFor_Empty } from './contracts'
-import { ProcessedTQ } from './utils'
+import { LoadingData } from './misc'
+import { ProcessedTQ } from './proposal'
 
 export { ProposalV1Beta1, ProposalV1 }
