+++ packages/types/nft.ts
@@ -1,15 +1,14 @@
 import { ComponentType, ReactNode, RefAttributes } from 'react'
 
-import { ChainId } from './chain'
+import { ChainId, WithChainId } from './chain'
 import {
   ButtonLinkProps,
   ButtonPopupSection,
-  LoadingDataWithError,
   ModalProps,
   StatefulEntityDisplayProps,
 } from './components'
 import { ContractInfoResponse } from './contracts/Cw721Base'
-import { WithChainId } from './state'
+import { LoadingDataWithError } from './misc'
 
 // Shape of type returned from Stargaze GraphQL indexer queries in
 // @dao-dao/state
