+++ packages/types/proposal.ts
@@ -1,6 +1,6 @@
+import { PolytoneConnection } from './chain'
 import { ProposalCardProps } from './components/ProposalCard'
 import { CosmosMsgFor_Empty } from './contracts'
-import { PolytoneConnection } from './utils'
 
 export type ProposalCreatedCardProps = Omit<
   ProposalCardProps,
@@ -78,3 +78,23 @@ export type DecodedIcaMsgNoMatch = {
 }
 
 export type DecodedIcaMsg = DecodedIcaMsgNoMatch | DecodedIcaMsgMatch
+
+export enum ProcessedTQType {
+  Majority,
+  Absolute,
+  Percent,
+}
+
+export type ProcessedTQ = { display: string } & (
+  | { type: ProcessedTQType.Majority }
+  | { type: ProcessedTQType.Absolute | ProcessedTQType.Percent; value: number }
+)
+
+export type ProcessedThresholdQuorum = {
+  threshold: ProcessedTQ
+  quorum?: ProcessedTQ
+}
+
+export type ProcessedQuorum = {
+  quorum: ProcessedTQ
+}
